+++ api/src/main/scala/dx/api/DxObject.scala
@@ -116,7 +116,9 @@ object DxObject {
   }
 }
 
-trait DxDataObject extends DxObject
+trait DxDataObject extends DxObject {
+  val project: Option[DxProject]
+}
 
 // Objects that can be run on the platform. These are apps, applets, and workflows.
 trait DxExecutable extends DxObject