+++ api/RELEASE_NOTES.md
@@ -3,9 +3,11 @@
 ## in develop
 
 * Unifies parsing of `choices`, `suggestions`, and `default` in `DxIOParameter`
-* No longer attempt to resolve files/projects/folders when parsing a `DxIOParameter`
-* Add tags field to `DxFileDescribe`
-* Add hidden field to `DxWorkflowDescribe`, `DxAppletDescribe`
+* No longer attempts to resolve files/projects/folders when parsing a `DxIOParameter`
+* Adds tags field to `DxFileDescribe`
+* Adds hidden field to `DxWorkflowDescribe`, `DxAppletDescribe`
+* `DxProject` now extends `DxObject` rather than `DxDataObject`
+* Adds `DxApi.dataObjectFromJson` method
 
 ## 0.6.0 (2021-07-12)